downloadProject.sh upd

+++ downloadProject.sh
@@ -3,39 +3,18 @@
 # Очистка консоли
 clear
 
-# Проверка наличия файла settings.conf
-if [[ ! -f settings.conf ]]; then
-    echo "Что-то пошло не так. Скачай актуальный firststart.sh скрипт по ссылке https://farm.simitka.io/farmjam-helper-installer"
+# Получение информации о последнем релизе из GitHub
+echo "Получение информации о последнем релизе..."
+release_info=$(curl -s https://api.github.com/repos/Simitka/FarmJam-Helper/releases)
+
+# Извлечение URL для последнего релиза и тега
+latest_release_url=$(echo "$release_info" | grep -o '"tarball_url": *"[^"]*' | head -1 | sed 's/"tarball_url": *"//')
+tag_name=$(echo "$release_info" | grep -o '"tag_name": *"[^"]*' | head -1 | sed 's/"tag_name": *"//')
+
+# Проверка на успешность получения URL и тега
+if [[ -z "$latest_release_url" || -z "$tag_name" ]]; then
+    echo "Не удалось получить информацию о последнем релизе. Попробуй позже."
     exit 1
-fi
-
-# Получение списка релизов из GitHub
-echo "Получение списка релизов..."
-releases=$(curl -s https://api.github.com/repos/Simitka/FarmJam-Helper/releases)
-
-# Поиск последнего релиза или pre-release
-latest_release=$(echo "$releases" | jq -r 'sort_by(.published_at) | reverse | .[] | select(.prerelease == false) | .tarball_url' | head -n 1)
-latest_pre_release=$(echo "$releases" | jq -r 'sort_by(.published_at) | reverse | .[] | select(.prerelease == true) | .tarball_url' | head -n 1)
-
-# Проверка на успешное получение URL
-if [[ -z "$latest_release" && -z "$latest_pre_release" ]]; then
-    echo "Не удалось получить информацию о релизах. Попробуй позже."
-    exit 1
-fi
-
-# Выбор новейшего релиза
-if [[ -n "$latest_release" && -n "$latest_pre_release" ]]; then
-    latest_release_date=$(curl -s $(echo "$releases" | jq -r --arg url "$latest_release" '.[] | select(.tarball_url == $url) | .published_at'))
-    latest_pre_release_date=$(curl -s $(echo "$releases" | jq -r --arg url "$latest_pre_release" '.[] | select(.tarball_url == $url) | .published_at'))
-
-    if [[ "$latest_pre_release_date" > "$latest_release_date" ]]; then
-        latest_release_url="$latest_pre_release"
-    else
-        latest_release_url="$latest_release"
-    fi
-else
-    latest_release_url="$latest_release"
-    [[ -z "$latest_release_url" ]] && latest_release_url="$latest_pre_release"
 fi
 
 # Добавление уникального тега к ссылке
@@ -54,7 +33,23 @@ if [[ $? -ne 0 ]]; then
     exit 1
 fi
 
-# Распаковка архива в текущий каталог
+# Удаление скриптов firstStart.sh и downloadProject.sh
+echo "Удаление старых скриптов..."
+if [[ -f firstStart.sh ]]; then
+    rm -f firstStart.sh
+    echo "Скрипт firstStart.sh удален."
+else
+    echo "Скрипт firstStart.sh не найден."
+fi
+
+if [[ -f downloadProject.sh ]]; then
+    rm -f downloadProject.sh
+    echo "Скрипт downloadProject.sh удален."
+else
+    echo "Скрипт downloadProject.sh не найден."
+fi
+
+# Распаковка архива в корень текущей директории
 echo "Распаковка архива..."
 tar -xzf latest_release.tar.gz --strip-components=1
 
@@ -65,11 +60,46 @@ if [[ $? -ne 0 ]]; then
 fi
 
 # Удаление архива
-rm latest_release.tar.gz
+echo "Удаление архива..."
+if [[ -f latest_release.tar.gz ]]; then
+    rm latest_release.tar.gz
+    echo "Архив latest_release.tar.gz удален."
+else
+    echo "Архив latest_release.tar.gz не найден."
+fi
 
-# Запрос нажатия любой кнопки перед запуском update.sh
+# Обновление файла settings.conf
+echo "Обновление файла settings.conf..."
+settings_file="settings.conf"
+if [[ -f $settings_file ]]; then
+    # Обновление существующего параметра
+    if grep -q "^tagVersion=" $settings_file; then
+        # Обновление параметра
+        sed -i '' "s/^tagVersion=.*/tagVersion=$tag_name/" $settings_file
+        if [[ $? -ne 0 ]]; then
+            echo "Ошибка при обновлении параметра tagVersion в $settings_file."
+            exit 1
+        fi
+    else
+        # Добавление нового параметра
+        echo "tagVersion=$tag_name" >> $settings_file
+        if [[ $? -ne 0 ]]; then
+            echo "Ошибка при добавлении параметра tagVersion в $settings_file."
+            exit 1
+        fi
+    fi
+else
+    # Создание нового файла и добавление параметра
+    echo "tagVersion=$tag_name" > $settings_file
+    if [[ $? -ne 0 ]]; then
+        echo "Ошибка при создании файла $settings_file."
+        exit 1
+    fi
+fi
+
+# Ожидание нажатия кнопки перед запуском update.sh
 echo "Для продолжения нажми любую кнопку..."
-read -n 1 -s
+read -n1 -s
 
 # Запуск скрипта update.sh
 if [[ -f update.sh ]]; then
